:art: config routes

+++ routes/collaborators.php
@@ -8,8 +8,8 @@ use App\Http\Controllers\Collaborators\CollaboratorsDestroyController;
 use Illuminate\Support\Facades\Route;
 
 // collaborators
-Route::get('collaborators', CollaboratorsIndexController::class)->name('index');
-Route::get('collaborators/{collaborator}', CollaboratorsShowController::class)->name('show');
-Route::post('collaborators', CollaboratorsStoreController::class)->name('store');
-Route::put('collaborators/{collaborator}', CollaboratorsUpdateController::class)->name('update');
-Route::delete('collaborators/{collaborator}', CollaboratorsDestroyController::class)->name('destroy');
+Route::get('collaborators', CollaboratorsIndexController::class)->name('index')->middleware(['role_or_permission:super-admin|collaborators.index']);
+Route::get('collaborators/{collaborator}', CollaboratorsShowController::class)->name('show')->middleware(['role_or_permission:super-admin|collaborators.show']);
+Route::post('collaborators', CollaboratorsStoreController::class)->name('store')->middleware(['role_or_permission:super-admin|collaborators.store']);
+Route::put('collaborators/{collaborator}', CollaboratorsUpdateController::class)->name('update')->middleware(['role_or_permission:super-admin|collaborators.update']);
+Route::delete('collaborators/{collaborator}', CollaboratorsDestroyController::class)->name('destroy')->middleware(['role_or_permission:super-admin|collaborators.destroy']);
