+++ routes/corporates.php
@@ -8,8 +8,8 @@ use App\Http\Controllers\Corporates\CorporateUpdateController;
 use Illuminate\Support\Facades\Route;
 
 // corporates
-Route::get('corporates', CorporateIndexController::class)->name('index');
-Route::get('corporates/{corporate}', CorporateShowController::class)->name('show');
-Route::post('corporates', CorporateStoreController::class)->name('store');
-Route::put('corporates/{corporate}', CorporateUpdateController::class)->name('update');
-Route::delete('corporates/{corporate}', CorporateDestroyController::class)->name('destroy');
+Route::get('corporates', CorporateIndexController::class)->name('index')->middleware(['role_or_permission:super-admin|corporate.index']);
+Route::get('corporates/{corporate}', CorporateShowController::class)->name('show')->middleware(['role_or_permission:super-admin|corporate.show']);
+Route::post('corporates', CorporateStoreController::class)->name('store')->middleware(['role_or_permission:super-admin|corporate.store']);
+Route::put('corporates/{corporate}', CorporateUpdateController::class)->name('update')->middleware(['role_or_permission:super-admin|corporate.update']);
+Route::delete('corporates/{corporate}', CorporateDestroyController::class)->name('destroy')->middleware(['role_or_permission:super-admin|corporate.destroy']);
